refactor: 이미지 업로더 UI, 로직 분리
- ImageUploader 컴포넌트 내부에서 UI와 미리보기 및 업로드 로직 분리
